Extract from templates internal subpackage the file updating machinery
