+++ pkg/scaffold/internal/templates/v2/crd/kustomization.go
@@ -22,7 +22,7 @@ import (
 
 	"sigs.k8s.io/kubebuilder/pkg/model/file"
 	"sigs.k8s.io/kubebuilder/pkg/model/resource"
-	"sigs.k8s.io/kubebuilder/pkg/scaffold/internal/templates/v2/internal"
+	"sigs.k8s.io/kubebuilder/pkg/scaffold/internal/machinery"
 )
 
 const (
@@ -62,7 +62,7 @@ func (f *Kustomization) Update() error {
 	kustomizeWebhookPatchCodeFragment := fmt.Sprintf("#- patches/webhook_in_%s.yaml\n", f.Resource.Plural)
 	kustomizeCAInjectionPatchCodeFragment := fmt.Sprintf("#- patches/cainjection_in_%s.yaml\n", f.Resource.Plural)
 
-	return internal.InsertStringsInFile(f.Path,
+	return machinery.InsertStringsInFile(f.Path,
 		map[string][]string{
 			kustomizeResourceScaffoldMarker:         {kustomizeResourceCodeFragment},
 			kustomizeWebhookPatchScaffoldMarker:     {kustomizeWebhookPatchCodeFragment},
